requirements.txt
alembic==0.9.6
blinker==1.4
click==6.7
docutils==0.14
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.3
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Geohash==1.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
MySQL-python==1.2.5
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.15
Werkzeug==0.13
xlrd==1.1.0
xlwt==1.3.0
